pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "pillow==11.0.0",
    "pypng==0.20220715.0",
    "pyqrcode==1.2.1",
]
